move PasswordDialog to src/widgets
