+++ src/widgets/PasswordDialog.js
